Updated html page skeleton and css file
Added the html basic templated included the grid container div
along with the 3 selections for the user to choose a color,
Create a grid, checkbox for rainbow trail option.
As part of css updated the styles for the mainbody and the
container for the grid and teh selection as a flex display
with column direction. the width of the contatiner is set as
800px x 800px within which everything is conatined.