Added the coloring part for the grid and completed project
Created a function named event listener which takes elemnt as
input and listens for mouserover event on the element and within
the callback changing the opacity and color of that element.
created a funciton which returns a random color value string in
an rgb format this is used to change the background color of the
div in callback function of event listener.